add loading indicator to advanced copy
