+++ src/components/Verse/AdvancedCopy/utils/copyText.ts
@@ -1,9 +1,5 @@
-import { getAdvancedCopyRawResult } from 'src/api';
-import { DEFAULT_MUSHAF } from 'src/utils/api';
 import { getVerseNumberFromKey } from 'src/utils/verse';
-import useSWRImmutable from 'swr/immutable';
 import clipboardCopy from 'clipboard-copy';
-import { makeAdvancedCopyUrl } from 'src/utils/apiPaths';
 
 /**
  * Validate the selected range start and end verse keys. The selection will be invalid in the following cases:
@@ -39,62 +35,13 @@ const validateRangeSelection = (
 };
 
 /**
- * Given these parameters, get the `text to be copied` from API
- */
-export const useTextToCopy = ({
-  verseKey,
-  showRangeOfVerses,
-  rangeStartVerse,
-  rangeEndVerse,
-  translations,
-  shouldCopyFootnotes,
-  shouldCopyText,
-}) => {
-  // by default the from and to will be the current verse.
-  let fromVerse = verseKey;
-  let toVerse = verseKey;
-  // if range of verse was selected
-  if (showRangeOfVerses) {
-    fromVerse = rangeStartVerse;
-    toVerse = rangeEndVerse;
-  }
-  // filter the translations
-  const toBeCopiedTranslations = Object.keys(translations).filter(
-    (translationId) => translations[translationId].shouldBeCopied === true,
-  );
-  const { data } = useSWRImmutable(
-    makeAdvancedCopyUrl({
-      raw: true,
-      from: fromVerse,
-      to: toVerse,
-      footnote: shouldCopyFootnotes,
-      ...(toBeCopiedTranslations.length > 0 && { translations: toBeCopiedTranslations.join(', ') }), // only include the translations when at least 1 translation has been selected.
-      ...(shouldCopyText && { mushaf: DEFAULT_MUSHAF }), // only include the Quranic text if the user chose to.
-    }),
-    () =>
-      getAdvancedCopyRawResult({
-        raw: true,
-        from: fromVerse,
-        to: toVerse,
-        footnote: shouldCopyFootnotes,
-        ...(toBeCopiedTranslations.length > 0 && {
-          translations: toBeCopiedTranslations.join(', '),
-        }), // only include the translations when at least 1 translation has been selected.
-        ...(shouldCopyText && { mushaf: DEFAULT_MUSHAF }), // only include the Quranic text if the user chose to.
-      }),
-  );
-
-  const result = data?.result;
-  return result || ''; // return empty text if the result is null.
-};
-
-/**
  * - Validate the range selection
  * - Copy the text to clipboard
  * - update the state if succeed or error
  * - set the error message if there's an error
  */
-export const copyText = ({
+
+const copyText = ({
   textToCopy,
   rangeStartVerse,
   rangeEndVerse,
@@ -116,3 +63,5 @@ export const copyText = ({
     setIsCopied(true);
   });
 };
+
+export default copyText;
