+++ src/components/dls/Modal/Action.module.scss
@@ -3,6 +3,8 @@
 .action {
   flex: 1;
   border: none;
+  display: flex;
+  justify-content: center;
   border-top: 1px solid var(--color-borders-hairline);
   font-size: var(--font-size-normal);
   background-color: inherit;
